+++ lib/arq/runnable.rb
@@ -16,9 +16,7 @@ module Arq
 
       validate_required_params
 
-      import_context
       val = run_block
-      export_context
 
       # If the block returned an array, attempt to run
       run_sequence(val) if val.is_a?(Array)
@@ -44,9 +42,11 @@ module Arq
     private
 
     def run_block
-      instance_eval(&@block)
-    rescue Arq::FailureError
-      nil
+      with_context do
+        instance_eval(&@block)
+      rescue Arq::FailureError
+        nil
+      end
     end
 
     def run_sequence(sequence)
@@ -60,20 +60,39 @@ module Arq
       end
     end
 
-    def import_context
+    # Runs the block with context parameters set as instance variables, then
+    # imports all context parameters and any new instance vars into the context.
+    # Returns the return value of the block.
+    def with_context
+      # Grab all current variables to know what is being returned
+      before_vars = instance_variables
+
+      # Load all context parameters into runnable instance
+      import_context_to_vars
+
+      # Run block
+      val = yield
+
+      # Grab all ctx + new vars
+      ctx_vars = instance_variables - before_vars
+
+      # Import instance vars into ctx
+      import_vars_to_context(ctx_vars)
+
+      # Return block value
+      val
+    end
+
+    def import_context_to_vars
       @ctx.each do |key, val|
         instance_variable_set(:"@#{key}", val)
       end
     end
 
-    def export_context
-      keys = [*@ctx.keys, *@returns]
-
-      keys.each do |key|
-        instance_key = :"@#{key}"
-        # Must check for existence since getting non-existent
-        # instance variables will return nil.
-        @ctx[key] = instance_variable_get(instance_key) if instance_variables.include?(instance_key)
+    def import_vars_to_context(vars)
+      vars.each do |var|
+        key = var[1..].to_sym
+        @ctx[key] = instance_variable_get(var)
       end
     end